wolfclaw_legacy/requirements.txt
streamlit==1.30.0
fastapi==0.103.2
uvicorn==0.23.2
httpx==0.24.1
pydantic==2.5.2
litellm==1.17.4
supabase==2.2.0
pywebview==4.4.1
paramiko==3.4.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
apscheduler==3.10.4
reportlab==4.0.8
backports.zoneinfo;python_version<"3.9"
typing-extensions>=4.5.0
